Update DNS Client
