+++ src/main/java/com/github/lexcorp3439/net/lab2/dns/model/MissingDomainException.java
@@ -1,0 +1,7 @@
+package com.github.lexcorp3439.net.lab2.dns.model;
+
+public class MissingDomainException extends RuntimeException {
+    public MissingDomainException() {
+        super();
+    }
+}
