+++ src/main/java/com/github/lexcorp3439/net/lab2/dns/model/rDataEntry.java
@@ -1,0 +1,9 @@
+package com.github.lexcorp3439.net.lab2.dns.model;
+
+import lombok.Data;
+
+@Data
+public class rDataEntry {
+    private int bytes;
+    private String domain;
+}
