--- src/main/java/lab2/dns/Main.java
@@ -1,12 +1,0 @@
-package lab2.dns;
-
-public class Main {
-    public static void main(String[] args) {
-        try {
-            DNSClient client = new DNSClient(args);
-            client.makeRequest();
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-    }
-}
