--- src/main/java/lab2/dns/model/MissingDomainException.java
@@ -1,7 +1,0 @@
-package lab2.dns.model;
-
-public class MissingDomainException extends RuntimeException {
-    public MissingDomainException() {
-        super();
-    }
-}
