--- src/main/java/lab2/dns/model/rDataEntry.java
@@ -1,9 +1,0 @@
-package lab2.dns.model;
-
-import lombok.Data;
-
-@Data
-public class rDataEntry {
-    private int bytes;
-    private String domain;
-}
